pyproject.toml
[project]
name = "vailresorts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-core>=1.11.2",
    "dbt-duckdb>=1.10.0",
    "iprogress>=0.4",
    "ipykernel>=7.1.0",
    "lazypredict",
    "matplotlib>=3.10.7",
    "optuna>=4.6.0",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "shap>=0.49.1",
    "xgboost>=3.1.2",
]

[tool.uv.sources]
lazypredict = { git = "https://github.com/shankarpandala/lazypredict.git" }
